requirements.txt
certifi==2023.7.22
charset-normalizer==3.2.0
config==0.5.1
decouple==0.0.7
idna==3.4
python-decouple==3.8
python-dotenv==1.0.0
requests==2.31.0
urllib3==2.0.4
